Strip unnecessary xmlns during merge
This should fix #121

The problem was when we emitted xml elements to be merged into the .targets file, the emit caused xmlns attributes to be added to each element.  While I think this is valid XML, it was causing errors when the packages were used with packages.config files.

This strips out all the xmlns attributes from the items being merged in, and should fix the issue.