Commit typeforwards from GenAPI directly
GenAPI is a very brittle tool, especially to have run xplat.  Since we don't anticipate doing these typeforwards again as Android Support 28.x is the last support library version that will be released, we are going to just manually commit the generated typeforwards this one time, and make it so that msbuild target won't actually run on build.  This resolves some headaches of trying to make something work that won't be needed (at least in the same way) again.