Remove AppBarLayout.BaseOnOffsetChangedListener binding
This one was causing issues in the java callable wrapper being generated since it wasn't being generated with the generic type argument on the java side.

Turns out, this might have caused issues in .NET bindings as well, so for now we are safe to remove it as there are still overloads anywhere `BaseOnOffsetChangedListener` is used that accept `OnOffsetChangedListener` and both interfaces contain the same method (as used in practice).