Switch to WebClient sync api
We were using `HttpClient` and calling `.Result` on one of the async methods that uses it which can result in a deadlock and might have been causing random locking in the VS IDE.  This switches to using `HttpWebRequest` and all synchronous API's to avoid this possible deadlock.  We also control the request timeout and set it to a lower 10 seconds as it's not a critical request we don't want to block the IDE too long.