Remove NuGet lookup from target
We initially tried to do a nice thing of, go find the latest version of android support libraries on nuget for a given api level if we wanted to recommend a version to use.  This proved to increase build times more often than we'd like.  Given that support libraries are winding down anyway, I don't think this is very relevant or important anymore.